Update source to include all changes upto android/atalk-ohos: Version: 00.09.000 (0009000); Release Date: 09/04/2025
Update source to include all changes upto android/atalk-ohos:
Version: 00.09.000 (0009000); Release Date: 09/04/2025

+++ entry/src/main/java/org/atalk/ohos/gui/settings/TimePreference.java
@@ -76,7 +76,7 @@ public class TimePreference extends DialogPreference
      * get the TimePreference persistent value the timePicker value update
      */
     public Long getPersistedValue() {
-        return (long) getPersistedLong(DEFAULT_VALUE);
+        return getPersistedLong(DEFAULT_VALUE);
     }
 
     @Override
